now we have boxes (also i changed everything to float64 because math library works better with it that way)

+++ main.go
@@ -7,8 +7,8 @@ import (
 const (
 	WindowWidth    int32   = 1000
 	WindowHeight   int32   = 600
-	PixelsPerMeter float32 = 10
-	MetersPerPixel float32 = 1.0 / PixelsPerMeter
+	PixelsPerMeter float64 = 10
+	MetersPerPixel float64 = 1.0 / PixelsPerMeter
 )
 
 func main() {
@@ -19,12 +19,12 @@ func main() {
 	for !rl.WindowShouldClose() {
 		game := NewCircleGame("Circle Game", 4, true)
 
-		previousTime := float32(rl.GetTime())
-		dt := float32(0.0)
+		previousTime := float64(rl.GetTime())
+		dt := float64(0.0)
 
 		for !rl.WindowShouldClose() {
-			dt = float32(rl.GetTime()) - previousTime
-			previousTime = float32(rl.GetTime())
+			dt = float64(rl.GetTime()) - previousTime
+			previousTime = float64(rl.GetTime())
 
 			game.UpdatePhysics(dt)
 			game.Draw()
